Merged all D-Bus interfaces into a single interface.
As the D-Bus interfaces used to be tightly integrated into the handler
classes, two d-bus interfaces had to be exposed. All the d-bus code has
now been moved into a single class and was separated from the handler
code. The interfaces exported by D-Bus have also been merged. Not only
is this cleaner than the previous approach but it also lets us easily
write unit tests for the D-Bus interface.