Added unit test suite for KDEd module.
Prepared unit tests for the KDE daemon module by adding an entry for a
static library to its CMakefile. This library is linked against all unit
tests for the daemon module as we can not link against the module's
shared object file directly.

Also added a mock object for the tablet handler and a stub for the
tablet service unit test. This allows us to test if the new library can
be linked successfully and if all Qt mocs are being generated by the
cmake macros.