Improved the X device event notifier.
Added an EventNotifier interface which is independent of the underlying
window system and adapted XDeviceEventNotifier to implement that
interface. Also renamed XDeviceEventNotifier to XEventNotifier as it
sends notifications about screen events, too.

The implemenation was separated from the interface so we can implement
a mock object for unit testing. Further we will be able to support
Wayland easily.