Added property adaptor mock template class.
Added a property adaptor template class which can be used as a mock for
any property that uses QString as its key type. This class will be
used by our unit tests as a mock for XinputAdaptor and
XsetwacomAdaptor. It lets us test the tablet handler and other central
classes without the need for a tablet being connected.