Added XInput device parser.
Added an XInput parser which uses the visitor pattern to allow easy
access to all input devices. This moves most of the code from X11Utils
into proper classes which is much cleaner and much more flexible. As
soon as all code is ported to the new parser, X11Utils will be removed.