Added X11 input visitor to find connected tablet devices.
Replaced more X11Utils code by adding a X11InputVisitor which is used to
find connected tablet devices. The visitor parses each input device
and checks if it is a tablet device. If a tablet is found, the tablet
information structure is filled with all details. This visitor is not
yet capable of distinguishing between multiple different tablet devices
being connected.