Disabled getInformation() method from D-Bus interface.
The method was only used by the applet anyway. Also the tablet
information structure will be extended soon when tablet detection is
enhanced, which makes it a less ideal candidate for a D-Bus interface
parameter. However the D-Bus getInformation() method might return soon
in a more basic version which is why it was not removed completely but
only disabled.