Reworked tablet detection and information.
Reworked the whole tablet detection process. It is now possible to
detect multiple tablet devices which are connected at the same time if
they are not of the same type. Still as the rest of the code is only
capable of handling one device, only the first tablet is actually used.
However it should now be impossible that devices (stylus/pad/...) of
different tablets are accidentally mixed up.

Also improved the amount of information which is actually queried from
the devices. This is necessary for libwacom support, as we do need i.e.
the full path to the input device.

Added a TabletFinder class which does all the hard work of detecting
as much information as possible of a device. At the moment it only
queries X11 for device information but soon it will query the filesystem
and the USB subsystem so we can support Intuos4 Leds.

Disabled non-working unit tests for now which have to be rewritten to
handle the new tablet information structures.