Renamed and moved some classes around.
Renamed some classes for a more consistent interface. Also moved some
classes from the common code to the kde daemon code as they are only
used by the daemon. They used to be in the common code for unit testing
but meanwhile we have a static unit testing library for the daemon code
as well.