Moved tablet information gathering to TabletFinder.
The tablet backend factory is no longer responsible for looking up
additional tablet information from the tablet database. Instead the
TabletFinder class is now not only responsible for finding tablets but
also for gathering all possible information. The backend factory is now
a simple factory as it is supposed to be. Based on the given tablet
information it creates a backend with all necessary property adaptors.