Support global shortcuts to switch screen mapping + TabletArea
The screen selected screen the user wants to use can now be switched
via shortcuts between "All screens"/"Screen 1"/"Screen 2"

Also the used TabletArea calibration can be set individually between
the selected screen mapping and will change accordingly. If the screen
is toggled the stylus/eraser will be set to Absolte mode automatically.

For the implementation the ScreenMapping property is now depricated
Instead the ScreenSpace can have the values
 * full
 * map0
 * map1
 * x y w h (for part of the screen mapping)

In full/map0/map1 mode the actually screen size as reported by qt will be
used automatically without looking into the config.
Whenever the screen mapping is changed the values are saved in the config
to be remembered for the next reconnect.

Also the Area value will be changed as copied from the
 * AreaMapFull
 * AreaMapPart
 * AreaMap0
 * AreaMap1

properties that store the tablet calibration configuration for each
screen mapping.

BUG: 313612
DIGEST: This patch makes working with multiple monitors a lot better