More button action selection changes.
Added ButtonActionDisplayWidget which is basically a KLineEdit but let's
us add some convenience features for the user. The user can now click
into the line edit widget which will open the button action selection.
Before he had to use the configuration button to assign an action. As
the configuration button is now no longer vital for the configuration
interface, it was changed to a flat button which looks much nicer.