Improved error widget.
The error widget for unsupported tablets and missing background
service now properly aligns to the top and looks more like the touchpad
error widget. This is to have a more consistent user experience.
The error widget will be changed to look exactly like the touchpad
error widget once I get to a computer without touchpad :-)