The return of the Touch configuration tab.
The touch configuration tab was reintroduced because the tracking mode
configuration had to be moved from the stylus tab to the tablet tab. The
tracking mode configuration needs to be in the tablet tab to make it
clear to the user that certain options are only available in absolute
mode.