Converted the screen area selection widget to MVP.
The ScreenAreaSelectionWidget was converted to a proper MVP model as it
will need access to many of the logic classes to do proper rotation
calculations. This code does not belong into a widget which is why the
widget was ported to a MVP model so we can properly separate the logic
from the view.