Inverted auto-rotation again.
Inverted the values for left/right auto-rotation again as this is the
value which should work with most configurations. All other users have
to use the auto-rotate inversion feature to correct for wrong events
sent by their display driver.

Auto rotation is tricky as it depends on the view point. If you rotate
the monitor to the left, then you have to rotate the screen canvas to
the right to get a correct screen orientation again. The same is true
for tablets. So wheter you have to rotate left or right depends on wether
your viewpoint is the monitor/tablet (hardware) or the screen/tablet
canvas.