Fixed a bug where unsupported buttons would override the settings for supported ones.
The KCM now makes sure that only buttons that the tablet device actually
supports are written to the configuration file. This prevents unsupported
buttons from overwriting the shortcuts of supported ones. If for instance
a valid button with number 3 gets mapped to the X11 button number 9, then
setting a shortcut for the unsupported button number 9 would actually
overwrite the shortcut of the mapped button 3.