Invalid settings are now pointed out to the user.
When the user set his tablet to relative mode and then applied a tablet
to screen mapping he could be left with settings which can not be
applied and without any hint why it is not working.

Instead of correcting the settings, a warning is displayed now to point
out configurations which can not be applied. Correcting the settings is
the wrong approach here because the user might actually want to stay in
relative mode and only switch to absolute mode with the global keyboard
shortcuts when he really needs the mapping.

Unfortunately some i18n strings had to be changed to make this clear.
Sorry for that!