Fixed screen mapping for touch devices.
Touch devices can now be mapped to a screen again. However it seems that
screen mapping for touch devices is only available in absolute mode, no
matter if the full touch area is selected or only a part of it. This
differs from the behaviour of the stylus which can be mapped to a screen
in relative mode as long as it is mapped to the full tablet area.

To make this clear to the user, the tablet area selection widget will
have to be updated again.