Fixed touch mapping warnings.
When using a touch device mapping which is only available in absolute
mode, a warning is now displayed. However this should be fixed
properly by disabling the controls. A warning is displayed for now as
it would take too many changes this close to release to fix it properly.
Also I'm not sure if this is a bug in the wacom driver.