Fixed bug which prevented special keys from being used as button shortcuts.
Added a reimplementation of KKeySequenceWidget which does not suppress
the shift modifier. This was necessary because the wacom driver treats
every key as keycode. Therefore, in order to get a "!", one has to use
the combination "key shift exclam" or "key shift 1". As the KDE
KKeySequenceWidget suppresses the shift modifier for known special keys,
it can not be used for Wacom shortcuts.

Also added some more special keys which are not yet supported in
xsetwacom 0.21 directly.