Support building with either xlib or xcb
Relying on xcb-xinput blocks successfully building the KDE wacomtablet
module on Debian and other distributions. The older KDE4 version, which
relied on XLib, can be restored to support most of the functionality.

This commit allows the build configuration to use the updated xcb-xinput
libraries if they are available on the system, or else fall back to
XLib. A message will be displayed at configuration time.