Add a button to launch kde_wacom_tabletfinder
Currenlty if tablet device is not in the tablet database, KCM will just
show an error claiming that device is not supported yet. However in most
cases you can use kde_wacom_tabletfinder to manually register the
device, which not everyone might be aware of.

Change the error message a bit and add a button to launch
kde_wacom_tabletfinder. This is suboptimal if there are multple tablets
connected, and one is detected, and another is not, but for now I'll
assume this rarely happens.