TabletDatabase: Add a version of `lookupTablet` which requires companyId
The existing `Wacom::TabletDatabase::lookupTablet` function does not require
a companyId and so may provide an unexpected match when provided with the
tabletId of a non-Wacom tablet which has the same ID as a Wacom tablet, or
the tablet ID of a Wacom tablet which isn't in the Wacom database but matches
a non-Wacom tablet. To allow for a more reliable match, this commit adds a
version of the lookupTablet function which requires a match on both the
provided companyID and the tabletID.

The first user of this new function is testlibwacomdata.cpp which needs to
accurately look up tablet data for comparison across the built-in and
libwacom databases.