libwacomwrapper: Calculate a better StatusLEDs value
The libwacom database assigns LEDs into groups that are each associated
with a ring or touchstrip control. The number of LEDs defines the number
of modes that the control makes available. The total number of StatusLEDs
is thus just the sum of the number of modes on each control. Note that
some devices have controls with multiple modes but no LEDs associated, so
simply counting *all* modes doesn't work; you must only count those modes
on controls returned by libwacom_get_status_leds().