dataengine: remove id from json file
The KPlugin documentation advises that C++ plugins should not set the
id field explicitly, as it will be set from the filename, and since
be10ddbf an explicit error message is shown indicating that this id
should be removed. The json file for the wacom tablet contains an id
field though, leading to an unnecessary error message every time the
dataengine is loaded (such as on plasmashell startup)

This change removes the unnecessary id field.